# Restore perf work.