… src/Compiler/Microsoft.NET.Sdk.Razor.SourceGenerators/RazorSourceGenerator.TagHelpers.cs
… src/Compiler/Microsoft.NET.Sdk.Razor.SourceGenerators/StaticCompilationTagHelperFeature.cs
… src/Compiler/Microsoft.NET.Sdk.Razor.SourceGenerators/StaticTagHelperFeature.cs
… src/Compiler/perf/Microsoft.AspNetCore.Razor.Microbenchmarks.Generator/AbstractBenchmark.cs
… src/Compiler/test/Microsoft.NET.Sdk.Razor.SourceGenerators.Tests/RazorSourceGeneratorTests.cs