Based on https://github.com/dotnet/razor/pull/8212

- Do an initial parse only once
- If a document has changed, re-write the tag helpers, recording which ones were used
- Re-run the tag helper re-write step, but check if the tag helpers have changed, and skip if we can