@chsienki 2023-07-20
Benchmarks show expected improvements, with very small regressions in scenarios that were already fast (becuase we're doing slightly more work to calculate e.g. the tag helpers used).

```
|                        Method |      Job | BuildConfiguration |       Mean |     Error |    StdDev | Ratio | RatioSD |      Gen0 |      Gen1 |   Allocated | Alloc Ratio |
|------------------------------ |--------- |------------------- |-----------:|----------:|----------:|------:|--------:|----------:|----------:|------------:|------------:|
|         Razor_Add_Independent | Baseline |      Release_Nuget | 110.111 ms | 1.8831 ms | 1.7615 ms |  1.00 |    0.00 | 4000.0000 |  500.0000 | 40476.39 KB |        1.00 |
|         Razor_Add_Independent |  Current |            Release |  33.367 ms | 0.6585 ms | 1.2844 ms |  0.32 |    0.01 |  333.3333 |  166.6667 |  4480.91 KB |        0.11 |
|                               |          |                    |            |           |           |       |         |           |           |             |             |
|        Razor_Edit_Independent | Baseline |      Release_Nuget |   9.412 ms | 0.1015 ms | 0.0949 ms |  1.00 |    0.00 |  343.7500 |  171.8750 |  3431.43 KB |        1.00 |
|        Razor_Edit_Independent |  Current |            Release |  10.464 ms | 0.2001 ms | 0.2054 ms |  1.11 |    0.02 |  328.1250 |  281.2500 |  3256.78 KB |        0.95 |
|                               |          |                    |            |           |           |       |         |           |           |             |             |
|      Razor_Remove_Independent | Baseline |      Release_Nuget | 103.775 ms | 2.0048 ms | 1.8753 ms |  1.00 |    0.00 | 4000.0000 |  500.0000 | 39472.93 KB |        1.00 |
|      Razor_Remove_Independent |  Current |            Release |  27.904 ms | 0.1345 ms | 0.1192 ms |  0.27 |    0.01 |  343.7500 |  281.2500 |  3677.78 KB |        0.09 |
|                               |          |                    |            |           |           |       |         |           |           |             |             |
| Razor_Edit_DependentIgnorable | Baseline |      Release_Nuget |   2.315 ms | 0.0102 ms | 0.0095 ms |  1.00 |    0.00 |   93.7500 |   15.6250 |   933.56 KB |        1.00 |
| Razor_Edit_DependentIgnorable |  Current |            Release |   3.110 ms | 0.0261 ms | 0.0244 ms |  1.34 |    0.01 |  109.3750 |   46.8750 |  1151.73 KB |        1.23 |
|                               |          |                    |            |           |           |       |         |           |           |             |             |
|          Razor_Edit_Dependent | Baseline |      Release_Nuget | 109.107 ms | 1.9709 ms | 1.7472 ms |  1.00 |    0.00 | 4000.0000 | 1000.0000 | 40163.54 KB |        1.00 |
|          Razor_Edit_Dependent |  Current |            Release |  37.089 ms | 0.6720 ms | 0.6286 ms |  0.34 |    0.01 |  400.0000 |  200.0000 |  4500.62 KB |        0.11 |
|                               |          |                    |            |           |           |       |         |           |           |             |             |
|        Razor_Remove_Dependent | Baseline |      Release_Nuget | 102.512 ms | 1.7309 ms | 1.6191 ms |  1.00 |    0.00 | 4000.0000 |  500.0000 | 39433.56 KB |        1.00 |
|        Razor_Remove_Dependent |  Current |            Release |  29.641 ms | 0.5633 ms | 0.5269 ms |  0.29 |    0.01 |  333.3333 |  166.6667 |  3651.84 KB |        0.09 |
```