@chsienki 2023-07-25
> Chris can you please manually test this compiler with some exemplar projects? ASP.NET, MudBlazor, etc?

@333fred Its been run against MudBlazor and AspNetCore here: https://dev.azure.com/dnceng/internal/_build/results?buildId=2229220&view=logs&j=426a675b-bf70-5b3e-81b7-b93f7cef5a5e&t=7023cd9d-f1a8-5290-89a7-d879caa30bad